+++ src/utils/__init__.py
@@ -2,7 +2,7 @@
 utils
 Samuel Koller
 Created: 18 October 2024
-Updated: 19 October 2024
+Updated: 22 October 2024
 
 Contains utilities for the program, such as configurations.
 """