+++ src/utils/constants.py
@@ -7,6 +7,6 @@ Updated: 21 October 2024
 Configurable constants to edit the default application.
 """
 
-TOP_FLOOR = 20  # Default: 20
+TOP_FLOOR = 20  # Default: 20, minimum of 20
 MAX_WEIGHT = 2000  # Default: 2000
 MAX_CAPACITY = 10  # Default: 10